+++ CHANGELOG.md
@@ -9,3 +9,5 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Added skeleton app
+- Added daos handlers
+- Added proposals handlers